+++ CHANGELOG.md
@@ -3,13 +3,13 @@
 ## 0.14.16 (tbd)
 
 - Fixed problem if provided pilet metadata is immutable (#486)
+- Upgraded `jest` dependencies to 26.x
 
 ## 0.14.15 (March 5, 2022)
 
 - Fixed issue with `get-dependency-map` of the `piral-debug-utils`
 - Added support for v0 and v1 pilets in `get-dependency-map`
 - Added support for object flags (e.g., `--vars`) in the initializers
-- Upgraded `jest` dependencies to 26.x.x
 
 ## 0.14.14 (March 3, 2022)